renamed plugin and added custom filenames for dumper
